+++ hy_xo/src/main/java/com/moxi/xo/service/impl/OptionBankServiceImpl.java
@@ -5,11 +5,14 @@ import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.moxi.base.enums.EOption;
+import com.moxi.base.enums.EPublicBank;
+import com.moxi.base.enums.EShareMode;
 import com.moxi.base.enums.EStatus;
 import com.moxi.base.serviceImpl.SuperServiceImpl;
 import com.moxi.utils.ResultUtil;
 import com.moxi.utils.StringUtils;
 import com.moxi.xo.entity.OptionSelect;
+import com.moxi.xo.entity.ProgramBank;
 import com.moxi.xo.global.MessageConf;
 import com.moxi.xo.global.SqlConf;
 import com.moxi.xo.global.SysConf;
@@ -41,13 +44,31 @@ public class OptionBankServiceImpl extends SuperServiceImpl<OptionBankMapper, Op
     @Override
     public IPage<OptionBank> getList(OptionBankVo vo) {
         QueryWrapper<OptionBank>wrapper=new QueryWrapper();
-        if(StringUtils.isNotEmpty(vo.getKeyword())){
-            wrapper.like(SqlConf.QUESTION_TITLE,vo.getKeyword());
+        if(vo.getIsPublic()== EPublicBank.SELF_BANk){
+            //若申请的是私人的题目集
+            wrapper=new QueryWrapper<OptionBank>(){{
+                //搜索关键字
+                if(StringUtils.isNotEmpty(vo.getKeyword())){
+                    like(SqlConf.QUESTION_TITLE,vo.getKeyword());
+                }
+                //PublisherId
+                eq(SqlConf.PUBLISHER_ID,vo.getPublisherId());
+                //降序排序
+                orderByDesc(SqlConf.CREATE_DATE);
+            }};
+        }else{
+            //公共题目集
+            wrapper=new QueryWrapper<OptionBank>(){{
+                //搜索关键字
+                if(StringUtils.isNotEmpty(vo.getKeyword())){
+                    like(SqlConf.QUESTION_TITLE,vo.getKeyword());
+                }
+                //shareMode
+                ge(SqlConf.SHARE_MODE,EShareMode.SHARE_ONLY_READ);
+                //降序排序
+                orderByDesc(SqlConf.CREATE_DATE);
+            }};
         }
-        if(vo.getOptionType()>0){
-            wrapper.eq(SqlConf.OPTION_TYPE,vo.getOptionType());
-        }
-        wrapper.orderByDesc(SqlConf.CREATE_DATE);
         Page<OptionBank> page=new Page<>(vo.getCurrentPage(),vo.getPageSize());
         IPage<OptionBank> list=optionBankService.page(page,wrapper);
         list.getRecords().forEach(x->{
@@ -67,32 +88,49 @@ public class OptionBankServiceImpl extends SuperServiceImpl<OptionBankMapper, Op
                     if(vo.getOptionType()!=EOption.JUDGE_CHOICE&&vo.getOptionVoList()!=null&&vo.getOptionVoList().size()>0){
                         selectLists= JSON.toJSONString(x.getOptionSelectList());
                     }
-                    return new OptionBank(x.getQuestionTitle(),selectLists,x.getOptionType(),x.getQuestionAnswer(),x.getPublisher(), EStatus.ENABLE,new Date(),new Date());
+                    return new OptionBank(x.getQuestionTitle(),selectLists,x.getOptionType(),x.getQuestionAnswer(),x.getPublisher(),x.getPublisherId(),x.getShareMode(),new Date(),new Date());
                 }).collect(Collectors.toList());
         optionBankService.saveBatch(optionList);
         return ResultUtil.result(SysConf.SUCCESS, MessageConf.INSERT_SUCCESS);
     }
 
     @Override
-    public String edit(OptionBankVo.OptionVo vo) {
+    public String edit(OptionBankVo.OptionVo vo, String userId) {
         OptionBank pre=optionBankService.getById(vo.getUid());
         if(pre==null)return ResultUtil.result(SysConf.ERROR,MessageConf.ENTITY_NOT_EXIST);
-        String questionSelects=null;
-        if(vo.getOptionSelectList()!=null&&vo.getOptionSelectList().size()>0){
-            questionSelects=JSON.toJSONString(vo.getOptionSelectList());
-        }else{
-            questionSelects=pre.getQuestionSelect();
+        boolean flag=false;
+        //验证权限
+        if(pre.getPublisherId().equals(userId)){
+            //如果创建该题目的用户
+            flag=true;
+        }else if(pre.getShareMode()== EShareMode.SHARE_READ_AND_WRITE){
+            //如果用户设置为可读写模式
+            flag=true;
         }
-        pre.UpdateOptionBank(vo.getQuestionTitle(),questionSelects, StringUtils.isNotBlank(vo.getQuestionAnswer()) ?vo.getQuestionAnswer():pre.getQuestionAnswer(),vo.getOptionType());
-        pre.updateById();
-        return ResultUtil.result(SysConf.SUCCESS,MessageConf.UPDATE_SUCCESS);
+        if(flag){
+            String questionSelects=null;
+            if(vo.getOptionSelectList()!=null&&vo.getOptionSelectList().size()>0){
+                questionSelects=JSON.toJSONString(vo.getOptionSelectList());
+            }else{
+                questionSelects=pre.getQuestionSelect();
+            }
+            pre.UpdateOptionBank(vo.getQuestionTitle(),questionSelects, StringUtils.isNotBlank(vo.getQuestionAnswer()) ?vo.getQuestionAnswer():pre.getQuestionAnswer(),vo.getOptionType(),vo.getShareMode());
+            pre.updateById();
+            return ResultUtil.result(SysConf.SUCCESS,MessageConf.UPDATE_SUCCESS);
+        }else{
+            return ResultUtil.result(SysConf.ERROR,MessageConf.INVALID_AUTH);
+        }
     }
 
     @Override
-    public String delete(String oid) {
+    public String delete(String userId, String oid) {
         OptionBank pre=optionBankService.getById(oid);
         if(pre==null)return ResultUtil.result(SysConf.ERROR,MessageConf.ENTITY_NOT_EXIST);
-        pre.deleteById();
-        return ResultUtil.result(SysConf.SUCCESS,MessageConf.DELETE_SUCCESS);
+        if(pre.getPublisherId().equals(userId)){
+            pre.deleteById();
+            return ResultUtil.result(SysConf.SUCCESS,MessageConf.DELETE_SUCCESS);
+        }else{
+            return ResultUtil.result(SysConf.ERROR,MessageConf.INVALID_AUTH);
+        }
     }
 }
